Лабораторная работа 7/pyproject.toml
[tool.poetry]
name = "лабораторная работа 7"
version = "0.1.0"
description = ""
authors = ["Alexander Volzhanin <alexandervolzhanin2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pillow = "^10.3.0"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
